Avance

+++ Weather/Models/City.cs
@@ -7,8 +7,8 @@ namespace Weather.Models
 {
     public class City
     {
-        public float TempMin { get; set; }
-        public float TempMax { get; set; }
+        public double TempMin { get; set; }
+        public double TempMax { get; set; }
         public string Icon { get; set; }
     }
 }
